refactor publishapi for easier import
